Validate inputs != outputs names
