+++ cloudslang-compiler/src/main/java/io/cloudslang/lang/compiler/modeller/SlangModellerImpl.java
@@ -8,12 +8,11 @@
  */
 package io.cloudslang.lang.compiler.modeller;
 
-import io.cloudslang.lang.compiler.Extension;
 import io.cloudslang.lang.compiler.SlangTextualKeys;
+import io.cloudslang.lang.compiler.Validator;
 import io.cloudslang.lang.compiler.modeller.result.ExecutableModellingResult;
 import io.cloudslang.lang.compiler.parser.model.ParsedSlang;
 import org.apache.commons.lang.Validate;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -28,6 +27,9 @@ public class SlangModellerImpl implements SlangModeller{
 
     @Autowired
     private ExecutableBuilder executableBuilder;
+
+    @Autowired
+    private Validator validator;
 
     @Override
     public ExecutableModellingResult createModel(ParsedSlang parsedSlang) {
@@ -57,24 +59,8 @@ public class SlangModellerImpl implements SlangModeller{
     private ExecutableModellingResult transformToExecutable(ParsedSlang parsedSlang, Map<String, Object> rawData) {
         String executableName = (String) rawData.get(SlangTextualKeys.EXECUTABLE_NAME_KEY);
         ExecutableModellingResult result = executableBuilder.transformToExecutable(parsedSlang, executableName, rawData);
-        validateFileName(executableName, parsedSlang, result);
+        validator.validateInputNamesDifferentFromOutputNames(result);
+        validator.validateFileName(executableName, parsedSlang, result);
         return result;
-    }
-
-    private void validateFileName(String executableName, ParsedSlang parsedSlang, ExecutableModellingResult result) {
-        String fileName = parsedSlang.getName();
-        Extension fileExtension = parsedSlang.getFileExtension();
-        if (StringUtils.isNotEmpty(executableName) && !executableName.equals(fileName)) {
-            if (fileExtension == null) {
-                result.getErrors().add(new IllegalArgumentException("Operation/Flow " + executableName +
-                        " is declared in a file named \"" + fileName + "\"," +
-                        " it should be declared in a file named \"" + executableName + "\" plus a valid " +
-                        "extension(" + Extension.getExtensionValuesAsString() + ") separated by \".\""));
-            } else {
-                result.getErrors().add(new IllegalArgumentException("Operation/Flow " + executableName +
-                        " is declared in a file named \"" + fileName + "." + fileExtension.getValue() + "\"" +
-                        ", it should be declared in a file named \"" + executableName + "." + fileExtension.getValue() + "\""));
-            }
-        }
     }
 }
