+++ cloudslang-compiler/src/main/java/io/cloudslang/lang/compiler/scorecompiler/ScoreCompilerImpl.java
@@ -10,19 +10,15 @@ package io.cloudslang.lang.compiler.scorecompiler;
 
 import ch.lambdaj.function.convert.Converter;
 import io.cloudslang.lang.compiler.SlangTextualKeys;
+import io.cloudslang.lang.compiler.Validator;
 import io.cloudslang.lang.compiler.modeller.DependenciesHelper;
 import io.cloudslang.lang.compiler.modeller.model.Executable;
 import io.cloudslang.lang.compiler.modeller.model.Flow;
 import io.cloudslang.lang.compiler.modeller.model.Operation;
-import io.cloudslang.lang.compiler.modeller.model.Step;
 import io.cloudslang.lang.entities.CompilationArtifact;
-import io.cloudslang.lang.entities.bindings.Argument;
-import io.cloudslang.lang.entities.bindings.Input;
-import io.cloudslang.lang.entities.bindings.Result;
 import io.cloudslang.score.api.ExecutionPlan;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Deque;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -30,7 +26,6 @@ import java.util.Map;
 import java.util.Set;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang.Validate;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -40,13 +35,16 @@ import static ch.lambdaj.Lambda.convertMap;
  * Created by stoneo on 2/2/2015.
  */
 @Component
-public class ScoreCompilerImpl implements ScoreCompiler{
+public class ScoreCompilerImpl implements ScoreCompiler {
 
     @Autowired
     private ExecutionPlanBuilder executionPlanBuilder;
 
     @Autowired
     private DependenciesHelper dependenciesHelper;
+
+    @Autowired
+    private Validator validator;
 
     @Override
     public CompilationArtifact compile(Executable executable, Set<Executable> path) {
@@ -101,151 +99,12 @@ public class ScoreCompilerImpl implements ScoreCompiler{
 
     private List<RuntimeException> validateModelWithDependencies(Executable executable, Map<String, Executable> dependencies) {
 
-        List<RuntimeException> errors = validateRequiredNonPrivateInputs(executable, dependencies);
+        List<RuntimeException> errors = validator.validateRequiredNonPrivateInputs(executable, dependencies);
 
         // Validate that all the steps of a flow have navigations for all the reference's results
-        errors.addAll(validateAllDependenciesResultsHaveMatchingNavigations(executable, dependencies));
+        errors.addAll(validator.validateAllDependenciesResultsHaveMatchingNavigations(executable, dependencies));
 
         return errors;
-    }
-
-    private List<RuntimeException> validateRequiredNonPrivateInputs(
-            Executable executable,
-            Map<String, Executable> filteredDependencies) {
-        Map<String, Executable> dependencies = new HashMap<>(filteredDependencies);
-        dependencies.put(executable.getId(), executable);
-        Set<Executable> verifiedExecutables = new HashSet<>();
-        return validateRequiredNonPrivateInputs(executable, dependencies, verifiedExecutables, new ArrayList<RuntimeException>());
-    }
-
-    private List<RuntimeException> validateRequiredNonPrivateInputs(
-            Executable executable,
-            Map<String, Executable> dependencies,
-            Set<Executable> verifiedExecutables,
-            List<RuntimeException> errors) {
-        //validate that all required & non private parameters with no default value of a reference are provided
-        if(!SlangTextualKeys.FLOW_TYPE.equals(executable.getType()) || verifiedExecutables.contains(executable)){
-            return errors;
-        }
-        verifiedExecutables.add(executable);
-
-        Flow flow = (Flow) executable;
-        Collection<Step> steps = flow.getWorkflow().getSteps();
-        Set<Executable> flowReferences = new HashSet<>();
-
-        for (Step step : steps) {
-            Executable reference = dependencies.get(step.getRefId());
-            List<String> mandatoryInputNames = getMandatoryInputNames(reference);
-            List<String> stepInputNames = getStepInputNames(step);
-            List<String> inputsNotWired = getInputsNotWired(mandatoryInputNames, stepInputNames);
-
-            try {
-                validateInputNamesEmpty(inputsNotWired, flow, step, reference);
-            } catch (RuntimeException e) {
-                errors.add(e);
-            }
-
-            flowReferences.add(reference);
-        }
-
-        for (Executable reference : flowReferences) {
-            validateRequiredNonPrivateInputs(reference, dependencies, verifiedExecutables, errors);
-        }
-        return errors;
-    }
-
-    private List<String> getMandatoryInputNames(Executable executable) {
-        List<String> inputNames = new ArrayList<>();
-        for (Input input : executable.getInputs()) {
-            if (!input.isPrivateInput() && input.isRequired() && input.getValue() == null) {
-                inputNames.add(input.getName());
-            }
-        }
-        return inputNames;
-    }
-
-    private List<String> getStepInputNames(Step step) {
-        List<String> inputNames = new ArrayList<>();
-        for (Argument argument : step.getArguments()) {
-            inputNames.add(argument.getName());
-        }
-        return inputNames;
-    }
-
-    private List<String> getInputsNotWired(List<String> mandatoryInputNames, List<String> stepInputNames) {
-        List<String> inputsNotWired = new ArrayList<>(mandatoryInputNames);
-        inputsNotWired.removeAll(stepInputNames);
-        return inputsNotWired;
-    }
-
-    private void validateInputNamesEmpty(List<String> inputsNotWired, Flow flow, Step step, Executable reference) {
-        Validate.isTrue(
-                CollectionUtils.isEmpty(inputsNotWired),
-                prepareErrorMessageValidateInputNamesEmpty(inputsNotWired, flow, step, reference)
-        );
-    }
-
-    private String prepareErrorMessageValidateInputNamesEmpty(List<String> inputsNotWired, Flow flow, Step step, Executable reference) {
-        StringBuilder inputsNotWiredBuilder = new StringBuilder();
-        for (String inputName : inputsNotWired) {
-            inputsNotWiredBuilder.append(inputName);
-            inputsNotWiredBuilder.append(", ");
-        }
-        String inputsNotWiredAsString = inputsNotWiredBuilder.toString();
-        if (StringUtils.isNotEmpty(inputsNotWiredAsString)) {
-            inputsNotWiredAsString = inputsNotWiredAsString.substring(0, inputsNotWiredAsString.length() - 2);
-        }
-        return "Cannot compile flow '" + flow.getId() +
-                "'. Step '" + step.getName() +
-                "' does not declare all the mandatory inputs of its reference." +
-                " The following inputs of '" + reference.getId() +
-                "' are not private, required and with no default value: " +
-                inputsNotWiredAsString + ".";
-    }
-
-    /**
-     * Validate that for all the steps in the flow, all results from the referenced operation or flow have matching navigations
-     * If the given {@link io.cloudslang.lang.compiler.modeller.model.Executable} is an operation, the method does nothing
-     * Throws {@link java.lang.IllegalArgumentException} if:
-     *      - Any reference of the executable is missing
-     *      - There is a missing navigation for one of the steps' references' results
-     *
-     * @param executable the flow to validate
-     * @param filteredDependencies a map holding for each reference name, its {@link io.cloudslang.lang.compiler.modeller.model.Executable} object
-     */
-    private List<RuntimeException> validateAllDependenciesResultsHaveMatchingNavigations(Executable executable, Map<String, Executable> filteredDependencies) {
-        List<RuntimeException> errors = new ArrayList<>();
-        if(executable.getType().equals(SlangTextualKeys.OPERATION_TYPE)){
-            return errors;
-        }
-        Flow flow = (Flow)executable;
-        Deque<Step> steps = flow.getWorkflow().getSteps();
-        for(Step step : steps){
-            List<Map<String, String>> stepNavigations = step.getNavigationStrings();
-            String refId = step.getRefId();
-            Executable reference = filteredDependencies.get(refId);
-            if (reference == null) {
-                errors.add(new IllegalArgumentException("Cannot compile flow: \'" + executable.getName() + "\' since for step: \'" + step.getName()
-                        + "\', the dependency: \'" + refId + "\' is missing."));
-            }
-            List<Result> refResults = reference.getResults();
-            for(Result result : refResults){
-                String resultName = result.getName();
-                if (!navigationListContainsKey(stepNavigations, resultName)){
-                    errors.add(new IllegalArgumentException("Cannot compile flow: \'" + executable.getName() +
-                            "\' since for step: '" + step.getName() + "\', the result \'" + resultName+
-                            "\' of its dependency: \'"+ refId + "\' has no matching navigation"));
-                }
-            }
-        }
-        return errors;
-    }
-
-    private boolean navigationListContainsKey(List<Map<String, String>> stepNavigations, String resultName) {
-        for (Map<String, String> map : stepNavigations) {
-            if (map.containsKey(resultName)) return true;
-        }
-        return false;
     }
 
     /**
