+++ cloudslang-tests/src/test/resources/yaml/functions/functions_test_op.sl
@@ -11,8 +11,8 @@ operation:
   name: functions_test_op
   inputs:
     - exist
-    - value_propagate:
-        default: ${ value_propagate + get_sp('propagate.op.input') }
+    - value_propagate_input:
+        default: ${ value_propagate_input + get_sp('propagate.op.input') }
         private: true
   python_action:
     script: |
@@ -30,7 +30,7 @@ operation:
     - output_7: ${get(    'i_don_exist',        get_sp('a.b.c.host')       )}
     - output_8: ${get_sp('a.b.c.i_don_exist', get_sp('a.b.c.host'))}
     - output_9: ${get_sp('a.b.c.null_value', 'default_str')}
-    - value_propagate: ${ value_propagate + get_sp('propagate.op.output') }
+    - value_propagate: ${ value_propagate_input + get_sp('propagate.op.output') }
     - output_10: ${get_sp('chars-b.c-hyphen')}
     - output_11: ${get_sp('chars-b.c-hyphen', 'default_str')}
     - output_12: ${get_sp("a.b.c.host")}
